requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-property==0.2.2
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
distro==1.9.0
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
lxml==5.2.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==2.0.0
openai==1.31.1
openapi-codec==1.3.2
openpyxl==3.1.4
packaging==24.0
pandas==2.2.2
prompt_toolkit==3.0.47
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
redis-lock==0.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
